Change a package name from `io` to `model`
