+++ core/src/main/java/hivemall/classifier/SoftConfideceWeightedUDTF.java
@@ -18,10 +18,10 @@
  */
 package hivemall.classifier;
 
-import hivemall.io.FeatureValue;
-import hivemall.io.IWeightValue;
-import hivemall.io.PredictionResult;
-import hivemall.io.WeightValue.WeightValueWithCovar;
+import hivemall.model.FeatureValue;
+import hivemall.model.IWeightValue;
+import hivemall.model.PredictionResult;
+import hivemall.model.WeightValue.WeightValueWithCovar;
 import hivemall.utils.math.StatsUtils;
 
 import javax.annotation.Nonnull;
