+++ core/src/main/java/hivemall/classifier/multiclass/MulticlassPerceptronUDTF.java
@@ -18,8 +18,8 @@
  */
 package hivemall.classifier.multiclass;
 
-import hivemall.io.FeatureValue;
-import hivemall.io.PredictionResult;
+import hivemall.model.FeatureValue;
+import hivemall.model.PredictionResult;
 
 import javax.annotation.Nonnull;
