+++ core/src/main/java/hivemall/ftvec/conv/ToDenseFeaturesUDF.java
@@ -17,7 +17,7 @@
  */
 package hivemall.ftvec.conv;
 
-import hivemall.io.FeatureValue;
+import hivemall.model.FeatureValue;
 
 import java.util.Arrays;
 import java.util.List;
