+++ core/src/main/java/hivemall/model/FeatureValue.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.io;
+package hivemall.model;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
