+++ core/src/main/java/hivemall/model/SpaceEfficientDenseModel.java
@@ -16,11 +16,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.io;
+package hivemall.model;
 
-import hivemall.io.WeightValue.WeightValueParamsF1;
-import hivemall.io.WeightValue.WeightValueParamsF2;
-import hivemall.io.WeightValue.WeightValueWithCovar;
+import hivemall.model.WeightValue.WeightValueParamsF1;
+import hivemall.model.WeightValue.WeightValueParamsF2;
+import hivemall.model.WeightValue.WeightValueWithCovar;
 import hivemall.utils.collections.IMapIterator;
 import hivemall.utils.hadoop.HiveUtils;
 import hivemall.utils.lang.Copyable;
