+++ core/src/main/java/hivemall/model/WeightValue.java
@@ -16,31 +16,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.io;
+package hivemall.model;
 
 import javax.annotation.Nonnegative;
 
-public class WeightValueWithClock implements IWeightValue {
+public class WeightValue implements IWeightValue {
 
     protected float value;
-    protected short clock;
-    protected byte deltaUpdates;
+    protected boolean touched;
 
-    public WeightValueWithClock(float value) {
-        this.value = value;
-        this.clock = 0;
-        this.deltaUpdates = 0;
+    public WeightValue() {}
+
+    public WeightValue(float weight) {
+        this(weight, true);
     }
 
-    public WeightValueWithClock(IWeightValue src) {
-        this.value = src.get();
-        if(src.isTouched()) {
-            this.clock = 1;
-            this.deltaUpdates = 1;
-        } else {
-            this.clock = 0;
-            this.deltaUpdates = 0;
-        }
+    public WeightValue(float weight, boolean touched) {
+        this.value = weight;
+        this.touched = touched;
     }
 
     @Override
@@ -53,22 +46,27 @@ public class WeightValueWithClock implements IWeightValue {
         throw new UnsupportedOperationException("getFloatParams(int) should not be called");
     }
 
+    @Override
     public final float get() {
         return value;
     }
 
+    @Override
     public final void set(float weight) {
         this.value = weight;
     }
 
+    @Override
     public boolean hasCovariance() {
         return false;
     }
 
+    @Override
     public float getCovariance() {
         throw new UnsupportedOperationException();
     }
 
+    @Override
     public void setCovariance(float cov) {
         throw new UnsupportedOperationException();
     }
@@ -91,68 +89,59 @@ public class WeightValueWithClock implements IWeightValue {
     /** 
      * @return whether touched in training or not
      */
+    @Override
     public final boolean isTouched() {
-        return deltaUpdates > 0;
+        return touched;
     }
 
     @Override
-    public void setTouched(boolean touched) {
-        throw new UnsupportedOperationException("WeightValueWithClock#setTouched should not be called");
+    public final void setTouched(boolean touched) {
+        this.touched = touched;
     }
 
+    @Override
     public final short getClock() {
-        return clock;
+        throw new UnsupportedOperationException();
     }
 
+    @Override
     public final void setClock(short clock) {
-        this.clock = clock;
+        throw new UnsupportedOperationException();
     }
 
+    @Override
     public final byte getDeltaUpdates() {
-        return deltaUpdates;
+        throw new UnsupportedOperationException();
     }
 
+    @Override
     public final void setDeltaUpdates(byte deltaUpdates) {
-        if(deltaUpdates < 0) {
-            throw new IllegalArgumentException("deltaUpdates is less than 0: " + deltaUpdates);
-        }
-        this.deltaUpdates = deltaUpdates;
+        throw new UnsupportedOperationException();
     }
 
     @Override
     public void copyTo(IWeightValue another) {
         another.set(value);
-        another.setClock(clock);
-        another.setDeltaUpdates(deltaUpdates);
+        another.setTouched(touched);
     }
 
     @Override
     public void copyFrom(IWeightValue another) {
         this.value = another.get();
-        this.clock = another.getClock();
-        this.deltaUpdates = another.getDeltaUpdates();
+        this.touched = another.isTouched();
     }
 
     @Override
     public String toString() {
-        return "WeightValueWithClock [value=" + value + ", clock=" + clock + ", deltaUpdates="
-                + deltaUpdates + "]";
+        return "WeightValue [value=" + value + "]";
     }
 
-    /**
-     * WeightValue with Sum of Squared Gradients
-     */
-    public static final class WeightValueParamsF1Clock extends WeightValueWithClock {
+    public static final class WeightValueParamsF1 extends WeightValue {
         private final float f1;
 
-        public WeightValueParamsF1Clock(float value, float f1) {
-            super(value);
+        public WeightValueParamsF1(float weight, float f1) {
+            super(weight);
             this.f1 = f1;
-        }
-
-        public WeightValueParamsF1Clock(IWeightValue src) {
-            super(src);
-            this.f1 = src.getFloatParams(1);
         }
 
         @Override
@@ -169,26 +158,23 @@ public class WeightValueWithClock implements IWeightValue {
         }
 
         @Override
-        public float getSumOfSquaredGradients() {
+        public final float getSumOfSquaredGradients() {
             return f1;
         }
 
     }
 
-    public static final class WeightValueParamsF2Clock extends WeightValueWithClock {
+    /**
+     * WeightValue with Sum of Squared Gradients
+     */
+    public static final class WeightValueParamsF2 extends WeightValue {
         private final float f1;
         private final float f2;
 
-        public WeightValueParamsF2Clock(float value, float f1, float f2) {
-            super(value);
+        public WeightValueParamsF2(float weight, float f1, float f2) {
+            super(weight);
             this.f1 = f1;
             this.f2 = f2;
-        }
-
-        public WeightValueParamsF2Clock(IWeightValue src) {
-            super(src);
-            this.f1 = src.getFloatParams(1);
-            this.f2 = src.getFloatParams(2);
         }
 
         @Override
@@ -207,12 +193,12 @@ public class WeightValueWithClock implements IWeightValue {
         }
 
         @Override
-        public float getSumOfSquaredGradients() {
+        public final float getSumOfSquaredGradients() {
             return f1;
         }
 
         @Override
-        public float getSumOfSquaredDeltaX() {
+        public final float getSumOfSquaredDeltaX() {
             return f2;
         }
 
@@ -223,19 +209,22 @@ public class WeightValueWithClock implements IWeightValue {
 
     }
 
-    public static final class WeightValueWithCovarClock extends WeightValueWithClock {
+    public static final class WeightValueWithCovar extends WeightValue {
         public static final float DEFAULT_COVAR = 1.f;
 
         private float covariance;
 
-        public WeightValueWithCovarClock(float value, float covar) {
-            super(value);
-            this.covariance = covar;
+        public WeightValueWithCovar() {
+            super();
         }
 
-        public WeightValueWithCovarClock(IWeightValue src) {
-            super(src);
-            this.covariance = src.getCovariance();
+        public WeightValueWithCovar(float weight, float covariance) {
+            this(weight, covariance, true);
+        }
+
+        public WeightValueWithCovar(float weight, float covariance, boolean touched) {
+            super(weight, touched);
+            this.covariance = covariance;
         }
 
         @Override
@@ -272,10 +261,8 @@ public class WeightValueWithClock implements IWeightValue {
 
         @Override
         public String toString() {
-            return "WeightValueWithCovar [value=" + value + ", clock=" + clock + ", deltaUpdates="
-                    + deltaUpdates + ", covariance=" + covariance + "]";
+            return "WeightValueWithCovar [value=" + value + ", covariance=" + covariance + "]";
         }
-
     }
 
 }
