+++ core/src/main/java/hivemall/regression/AdaGradUDTF.java
@@ -19,9 +19,9 @@
 package hivemall.regression;
 
 import hivemall.common.LossFunctions;
-import hivemall.io.FeatureValue;
-import hivemall.io.IWeightValue;
-import hivemall.io.WeightValue.WeightValueParamsF1;
+import hivemall.model.FeatureValue;
+import hivemall.model.IWeightValue;
+import hivemall.model.WeightValue.WeightValueParamsF1;
 import hivemall.utils.lang.Primitives;
 
 import javax.annotation.Nonnull;
