+++ core/src/main/java/hivemall/regression/PassiveAggressiveRegressionUDTF.java
@@ -20,8 +20,8 @@ package hivemall.regression;
 
 import hivemall.common.LossFunctions;
 import hivemall.common.OnlineVariance;
-import hivemall.io.FeatureValue;
-import hivemall.io.PredictionResult;
+import hivemall.model.FeatureValue;
+import hivemall.model.PredictionResult;
 
 import javax.annotation.Nonnull;
